Add base protection proto

+++ conda/cli/install2.py
@@ -237,6 +237,13 @@ def _prefix_dependent_checks(environment: Environment, command: str, args: Names
         if args.clone:
             _check_clone(args)
     elif environment.exists():
+        # Check if we are editing 'base'
+        if not args.break_base_env and paths_equal(environment.prefix, context.root_prefix):
+            raise CondaValueError(
+                "The target prefix is the base prefix. This is discouraged. If you really want to "
+                "modify the 'base' environment, use the --break-base-env flag at your own risk."
+            )
+
         delete_trash(environment.prefix)
         # TODO: If we changed the Solver logic, this merged environment could
         # hold all the information required to simply invoke the solution.
